Fixed Android Lint errors in the main client app DaisyBookLister.
There are numerous warnings still to address. Some of these are in the dependent opensource actionbarsherlock library which need to be carefully considered if we want to accept updates to that project without incurring ongoing headaches involved with patching our changes.

Note: I ran Android Lint mainly from the command line, as follows:
  lint --quiet --nowarn .

When I run Android Lint in Eclipse I seem to get a different set of results. For instance, it includes the errors for actionbarsherlock (maybe I need to find a way to configure it better in Eclipse?).